oh my god it actually just works
